+++ src/Concerns/InteractsWithLogger.php
@@ -12,9 +12,12 @@ trait InteractsWithLogger
 	 */
 	protected function fakesLogger()
 	{
-		return $this->swap('logger', function (Container $c) {
+		$this->swap('logger', function (Container $c) {
 			return new Logger($this->app);
 		});
+
+		// resolve out of the container - swap() hands back the closure, not the instance
+		return $this->getLoggerFake();
 	}
 
 	/**
@@ -66,7 +69,8 @@ trait InteractsWithLogger
     protected function assertActionLogged($type, $callback = null)
     {
         if (is_numeric($callback)) {
-            return $this->assertActionLoggedTimes($callback);
+            $this->assertActionLoggedTimes($type, $callback);
+            return;
         }
 
 	    $loggedActions = $this->loggedActions($type, $callback);
@@ -90,7 +94,8 @@ trait InteractsWithLogger
     protected function assertChangeLogged($type, $callback = null)
     {
         if (is_numeric($callback)) {
-            return $this->assertChangeLoggedTimes($callback);
+            $this->assertChangeLoggedTimes($type, $callback);
+            return;
         }
 
 	    $loggedChanges = $this->loggedChanges($type, $callback);