+++ src/Job/Manager.php
@@ -1,6 +1,7 @@
 <?php namespace Hampel\Testing\Job;
 
 use XF\Job\JobParams;
+use XF\Job\JobResult;
 use XF\Job\Manager as BaseManager;
 
 class Manager extends BaseManager
@@ -29,6 +30,7 @@ class Manager extends BaseManager
 	 */
 	public function runByIds(array $ids, $maxRunTime)
 	{
+		return [];
 	}
 
 	/**
@@ -54,7 +56,7 @@ class Manager extends BaseManager
 	 * @param array $job
 	 * @param int $maxRunTime
 	 *
-	 * @return JobResult
+	 * @return JobResult|null
 	 */
 	public function runJobEntry(array $job, $maxRunTime)
 	{